move eagle prepare_for_decode to free fn; dissolve EagleDraftInputV2Mixin
